+++ code/ui/interaction/menu/filter/filter-menu.tsx
@@ -13,6 +13,7 @@ import { getInitialDateFromLifecycleStage } from "@/utils/client-utils";
 import { ColumnFilter } from "@tanstack/react-table";
 import { Dispatch, SetStateAction, useState } from "react";
 import { FieldValues } from "react-hook-form";
+import { FunnelX, ListFilter } from "lucide-react";
 
 interface FilterMenuProps {
     hasNoActiveFilters: boolean;
@@ -60,7 +61,7 @@ export default function FilterMenu(props: Readonly<FilterMenuProps>) {
         placement="bottom"
         draggable
         bottomSheet
-        leftIcon="filter_list"
+        leftIcon={ListFilter}
         variant={props.hasNoActiveFilters ? "outline" : "secondary"}
         isOpen={isMenuOpen}
         setIsOpen={setIsOpen}
@@ -111,10 +112,9 @@ export default function FilterMenu(props: Readonly<FilterMenuProps>) {
         </section>
         <footer className="shrink-0 -mx-2 border-t border-border pt-2 px-3">
             <Button
-                leftIcon="filter_list_off"
+                leftIcon={FunnelX}
                 label={dict.action.clearAllFilters}
                 aria-label={dict.action.clearAllFilters}
-                iconSize="medium"
                 disabled={props.hasNoActiveFilters || !isConnected}
                 onClick={() => props.resetFilters()}
                 variant="outline"